Fix and restructure tests
